added the info interface to guild model
