+++ src/models/command.ts
@@ -8,10 +8,3 @@ export interface Command {
   description: string;
   response: Response | ((msg: Message, guildConfig: GuildConfig) => PromiseLike<Response | void>)
 }
-
-export interface Info {
-  name: string,
-  aliases: string[],
-  description: string;
-  reply: Command[]
-}
