+++ src/models/guild.ts
@@ -1,4 +1,4 @@
-import { Command, Info } from './command';
+import { Command } from './command';
 
 export interface GuildConfig {
   _id: string,
@@ -7,6 +7,13 @@ export interface GuildConfig {
   info: Info,
   commands: Command[],
   notes: Record<string, string[]>
+}
+
+export interface Info {
+  name: string,
+  aliases: string[],
+  description: string;
+  reply: Command[]
 }
 
 export interface CouresChannel
